+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Preserve leading indentation in fixed doc comments (#821).
 * Add `--verbose` to hide advanced options in `--help` output.
 * Split outer nested control flow elements (#869).
+* Don't add unneeded splits on if elements near comments (#888).
 
 # 1.3.3