+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 * Add `--fix-single-cascade-statements`.
 * Correctly handle `var` in `--fix-function-typedefs` (#826).
 * Preserve leading indentation in fixed doc comments (#821).
+* Add `--verbose` to hide advanced options in `--help` output
 * Add `--verbose` to hide advanced options in `--help` output.
 * Split outer nested control flow elements (#869).
+* Always place a blank line after script tags (#782).
 * Don't add unneeded splits on if elements near comments (#888).
 
 # 1.3.3